requirements.txt
streamlit==1.37.0
pandas==2.2.2
plotly==5.23.0
xrpl-py==3.0.0
requests==2.32.3
